# yoxa/openapi/calculate-contribution.openapi.yaml
openapi: 3.1.0
info:
  title: Calculate Expected Contribution
  version: 1.0.0
servers:
  - url: https://uzvqxpfdxwckhtvvjkzo.supabase.co
paths:
  /functions/v1/calculate-contribution:
    post:
      operationId: calculateExpectedContribution
      summary: Calculate expected patient and insurer contribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [case_id, case_version, currency, eligible_line_items, non_payable_items, contracted_package_rate, co_payment_rate, deductible_amount]
              properties:
                case_id:
                  type: string
                case_version:
                  type: integer
                currency:
                  type: string
                eligible_line_items:
                  type: array
                  items:
                    type: number
                non_payable_items:
                  type: array
                  items:
                    type: number
                contracted_package_rate:
                  type: number
                applicable_sub_limit:
                  type: number
                  nullable: true
                co_payment_rate:
                  type: number
                deductible_amount:
                  type: number
      responses:
        '200':
          description: Contribution calculation complete
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    additionalProperties: false
                    required: [case_id, currency, expected_insurer_contribution, expected_patient_contribution, status]
                    properties:
                      case_id:
                        type: string
                      currency:
                        type: string
                      expected_insurer_contribution:
                        type: number
                      expected_patient_contribution:
                        type: number
                      status:
                        type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []